add ticket template refactor.
