+++ client/src/components/pages/Store/Settings/TicketTemplate/Footer.jsx
