+++ client/src/components/pages/Store/Settings/TicketTemplate/List.jsx
