+++ client/src/components/pages/Store/Settings/TicketTemplate/Preview.jsx
