+++ client/src/components/pages/Store/Settings/TicketTemplate/VariablePicker.jsx
