+++ client/src/components/pages/Store/Settings/TicketTemplate/defaults.js
