+++ client/src/components/pages/Store/Settings/TicketTemplate/sampleData.js
