+++ client/src/components/pages/Store/Settings/__tests__/PrinterSettingsCard.test.jsx
@@ -56,7 +56,7 @@ jest.mock("../Printer/PrinterConfigRow", () => ({
   ),
 }));
 
-jest.mock("../TicketTemplate/TicketTemplatesModal", () => ({
+jest.mock("../TicketTemplate/Modal", () => ({
   TicketTemplatesModal: ({ isOpen }) => (isOpen ? <div>templates-modal</div> : null),
 }));
